+++ libraries/AtomicWeight/AtomicWeight.cpp
@@ -2,7 +2,7 @@
 //    FILE: AtomicWeight.cpp
 //  AUTHOR: Rob Tillaart
 //    DATE: 2022-03-09
-// VERSION: 0.2.2
+// VERSION: 0.3.0
 // PURPOSE: Arduino library for atomic weights
 //     URL: https://github.com/RobTillaart/AtomicWeight
 
@@ -98,6 +98,17 @@ float PTOE::massPercentage(const char * formula, const char * abbrev)
 }
 
 
+float PTOE::weightEV(const uint8_t element)
+{
+  return weight(element) * DALTON2EV;
+}
+
+
+float PTOE::weightEV(const char * formula, const char * abbrev)
+{
+  return weight(formula, abbrev) * DALTON2EV;
+}
+
 
 ////////////////////////////////////////////////////////////////
 //
@@ -112,6 +123,12 @@ float PTOE::moles2grams(const char * formula, float moles)
 float PTOE::grams2moles(const char * formula, float grams)
 {
   return grams / weight(formula);
+}
+
+
+float PTOE::dalton2EV(float Dalton)
+{
+  return Dalton * DALTON2EV;
 }
 
 
@@ -245,7 +262,7 @@ float PTOE::_weight(const char separator, const char * abbrev)
   float sum = 0;
   float w   = 0;
   char elem[3] = { 0, 0, 0 };
-  uint32_t count = 0;
+  float count = 0;
 
   while (*p != separator)
   {
@@ -285,6 +302,17 @@ float PTOE::_weight(const char separator, const char * abbrev)
       count = count * 10 + (*p - '0');
       p++;
     }
+    if (*p == '.')  //  floating point numbers
+    {
+      p++;  //  skip decimal .
+      float divider = 0.1;
+      while (isdigit(*p))
+      {
+        count += (*p - '0') * divider;
+        divider *= 0.1;
+        p++;
+      }
+    }
     //  correct for no digits
     if (count == 0) count = 1;
 
@@ -297,11 +325,11 @@ float PTOE::_weight(const char separator, const char * abbrev)
 }
 
 
-uint32_t PTOE::_count(const char separator, const char * abbrev)
+float PTOE::_count(const char separator, const char * abbrev)
 {
-  uint32_t sum = 0;
+  float sum = 0;
   char elem[3] = { 0, 0, 0 };
-  uint32_t count = 0;
+  float count = 0;
   int w = 0;
 
   while (*p != separator)
@@ -341,6 +369,17 @@ uint32_t PTOE::_count(const char separator, const char * abbrev)
     {
       count = count * 10 + (*p - '0');
       p++;
+    }
+    if (*p == '.')  //  floating point numbers
+    {
+      p++;  //  skip decimal .
+      float divider = 0.1;
+      while (isdigit(*p))
+      {
+        count += (*p - '0') * divider;
+        divider *= 0.1;
+        p++;
+      }
     }
     //  correct for no digits
     if (count == 0) count = 1;