+++ libraries/AtomicWeight/test/unit_test_001.cpp
@@ -52,10 +52,10 @@ unittest(test_constants)
   assertEqualFloat(1.66053907e-24, DALTON, 1e-30);
   assertEqualFloat(1.0, DALTON * AVOGADRO, 1e-5);
 
-  assertEqualFloat(1.602176565e-19, ELEKTRON_VOLT_JOULE, 1e-5);
-  assertEqualFloat(1.7826619e-39, ELEKTRON_VOLT_GRAM, 1e-5);
-  assertEqualFloat(DALTON / ELEKTRON_VOLT_GRAM, DALTON_EV, 1e-5);
-  assertEqualFloat(DALTON / ELEKTRON_VOLT_JOULE, DALTON_JOULE, 1e-5);
+  assertEqualFloat(1.6726231E-24, PROTON_WEIGHT, 1e-30);
+  assertEqualFloat(1.6749286E-24, NEUTRON_WEIGHT, 1e-30);
+  assertEqualFloat(9.10938356e-28, ELECTRON_WEIGHT, 1e-34);
+  assertEqualFloat(931494697.25613, DALTON2EV, 1e2);
 }